Genomica/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# API Documentation
drf-spectacular==0.27.0

# Filtering and searching
django-filter==23.5

# CORS headers
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Utilities
pytz==2023.3
